Added support for default node.
Addresses issue #24.

Node is created on startup by using a Java agent and the :java-agents lein
configuration parameter.

Note, however, since this is part of the same source, if building it for the
first time, you have to build initial with the :java-agents option commented
out, and then enable it once it's build.

Also done in this branch:

* Added splash screen support (addresses issue #23)
* Initial work on node functions with decreased arity